Allow xml to not be order dependent
Our old version of testing required all the xml elements to be in a
specific order, and that is not required by the spec. I pulled in an
older xml equivalence gem that hasn't been updated in a while, but seems
to do the job really well. It has an option for ignoring attributes
which we want for our ids.

Amos King @adkron <amos@binarynoggin.com>