Add minimal slide with one text box
This was the minimal slide I could produce that would have a bax in the
middle that said hello world.

This will only add one spot of text in the upper left. It also has a
fixed size. We will need to decide on size based on text.

Amos King @adkron <amos@binarynoggin.com>