Adds the ability to have more than one text box
This should lead us to be able to have more than one shape that is more
than text. All the groundwork is in place. One thing that I noticed is
that a name of a shape doesn't have to be unique. Since we have the
unique id I think that having the name be representative of the object
is a good step in helping to debug.

Amos King @adkron <amos@binarynoggin.com>