Add image to slide
This was a squashed commit and I kept the original commit messages
because I think they shed some light on the process.

Amos King @adkron <amos@binarynoggin.com>

------------------------------

This is the smallest amount of change I could get to add a picture to a
slide. I put text on the slide too because I think that is a normal use
case and they should work flawlessly together.

Amos King @adkron <amos@binarynoggin.com>

Add image xml

This adds all xml needed for an image except for content type.

Amos King @adkron <amos@binarynoggin.com>

Add image file and content type

All the image content types have the same name as the extension except
for the short hand `jpg` file path. This is why I only made a single entry
in the `TYPES` and allow the rest to default. This gives us a nice
extension point for the future. It would be nice to detect the mime type
instead of relying on the extension except that we can't reliably do
that on every OS to my knowledge.

Amos King @adkron <amos@binarynoggin.com>