Only allow typeface and font_color to be set on initialize
At first I was thinking that we might want these properties to be
overridden. I chose to make them only setable on initialization because
if this text piece appears on multiple slides we don't want something to
change the color out from under us. If we want all of them to change the
code should change on creating the text block.

Amos King @adkron <amos@binarynoggin.com>